Off-by-one error, print area optional, manager function
Corrected the off-by-one error so the user can input the worksheet
number starting with 1 and the program will correct for it.

Originally, defining the print area was not a option, it was required.
It is not an optional definition, if it is not defined the print area of
each worksheet will be used as is. This will need additional work to
specify the paper size.

Defined a manager function for the user to more easily pass arguments to
and run the program. This will need to be refactored; in it's current
state is uses the print_multiple_ws function and does not give the user
a choice to only print one worksheet.